Created Coding- Grading Scale
Created code to prompts user to enter a score. Uses ‘try’ and ‘except’ to either convert the entered string to a float number or request a convertible number. Use ‘if’ to see if the number is within the grading scale range and if not, requests a number in the range. Uses ‘if’ and ‘elif’ to find and print the correct letter grade.